+++ views/mail/editEntry.php
@@ -24,7 +24,7 @@ use yii\bootstrap\ActiveForm;
 
 <?php $form = ActiveForm::begin() ?>
 <div class="modal-body mail-edit-message">
-    <?= $form->field($entry, 'content')->widget(
+    <?= $form->field($entry, 'entry_content')->widget(
         MailRichtextEditor::class, [
         'placeholder' => Yii::t('MailModule.base', 'Edit message...')])->label(false) ?>