+++ widgets/ConversationEntryLinks.php
@@ -1,0 +1,40 @@
+<?php
+/**
+ * @link https://www.humhub.org/
+ * @copyright Copyright (c) 2023 HumHub GmbH & Co. KG
+ * @license https://www.humhub.com/licences
+ */
+
+namespace humhub\modules\mail\widgets;
+
+use humhub\modules\like\widgets\LikeLink;
+use humhub\modules\mail\models\MessageEntry;
+use humhub\widgets\BaseStack;
+
+class ConversationEntryLinks extends BaseStack
+{
+    /**
+     * @var MessageEntry
+     */
+    public $object = null;
+
+    /**
+     * @inheritdoc
+     */
+    public $seperator = '&nbsp;&middot;&nbsp;';
+
+    /**
+     * @inheritdoc
+     */
+    public function init()
+    {
+        $this->initDefaultWidgets();
+        parent::init();
+    }
+
+    public function initDefaultWidgets()
+    {
+        $this->addWidget(LikeLink::class, ['object' => $this->object], ['sortOrder' => 200]);
+    }
+
+}